unit path movement
This allows units to make multiple moves.  You use the set_path method
with a destination, and it will assign the unit a list of moves to
follow to get to that destination (currently it just goes in a straight
line, eventually should find the lowest cost path).
render_unit_movement will follow that path one at a time.

Also converts some of Tucker's degenerate space-indentation to proper
tab-indentation :).